requirements.txt
visualtorch>=0.1.0
torchrl>=0.2.0
torchsummary>=1.5.1
tensordict>=0.2.1
easyocr>=1.4
pandas>=1.4.1
matplotlib>=3.5.1
scikit-learn>=1.0
seaborn>=0.11.2
tqdm>=4.62.3
